chore: refactor to reactive controller
